8/29 수정사항 반영

+++ src/main/java/com/dsqd/amc/linkedmo/controller/SubscribeController.java
@@ -15,6 +15,7 @@ import com.dsqd.amc.linkedmo.model.Data;
 import com.dsqd.amc.linkedmo.model.Subscribe;
 import com.dsqd.amc.linkedmo.service.DataService;
 import com.dsqd.amc.linkedmo.service.SubscribeService;
+import com.dsqd.amc.linkedmo.util.InterfaceManager;
 
 import net.minidev.json.JSONArray;
 import net.minidev.json.JSONObject;
@@ -23,7 +24,8 @@ import net.minidev.json.JSONValue;
 public class SubscribeController {
 	private static final Logger logger = LoggerFactory.getLogger(DataController.class);
 	private SubscribeService service = new SubscribeService();
-
+	private InterfaceManager itfMgr = InterfaceManager.getInstance();
+	
 	public SubscribeController() {
 		setupEndpoints();
 	}
@@ -89,34 +91,58 @@ public class SubscribeController {
 						
 						//기존에 동일한 휴대전화번호가 있는지 확인함 
 						List <Subscribe> checkDup = service.getSubscribeByMobileno(data);
+						
+						//오늘 가입한 적이 있는 휴대전화번호인지 확인함
+						List <Subscribe> todayMobileno = service.getTodaySubscribeByMobileno(data);
+						
 						if (checkDup.size() > 0) {
 							logger.warn("Already Subscribed Mobile: {}", data.getMobileno());
 							code = 901;
-							msg = "이미 가입된 전화번호입니다.";
-						} else {
+							msg = "이미 가입된 전화번호입니다.[901]";
+							
+						} else if (todayMobileno.size() > 0) { // 오늘 가입한적이 있음 
+							code = 902;
+							msg = "입력하신 전화번호는 해지관련 전산처리 중으로 내일 가입이 가능합니다.[902]";
+							
+						} else { // 정상 가입 
 							
 							// 통신사로 부가서비스 가입요청
 							
 							// 부가서비스 제공사로 가입요청
+							String method = "POST";
+							String uri = "/api/v1.0/linksafe/subscribe";
+							JSONObject jsonParam = new JSONObject();
+							jsonParam.put("spcode", data.getSpcode());
+							jsonParam.put("mobileno", data.getMobileno());
 							
+							JSONObject itfJSON = (JSONObject) JSONValue.parse(itfMgr.sendRequest(method, uri, jsonParam.toJSONString()));
+							String resultCode = itfJSON.getAsString("code");
+							String resultMsg = itfJSON.getAsString("msg");
 							// 아무 문제가 없으면 DB에 저장함 
-							try {
-								service.insertSubscribe(data);
-								logger.info("Subscribe inserted: {}", data);
-								res.status(201);
-								code = 200;
-								msg = "정상 가입";
-								
-							} catch (Exception e) {
-								e.printStackTrace();
-								code = 999;
-								msg = "가입 중 오류가 발생하였습니다.";
+							
+							if ("200".equals(resultCode)) {
+								try {
+									service.insertSubscribe(data);
+									logger.info("Subscribe inserted: {}", jsonParam.toJSONString());
+									res.status(201);
+									code = 200;
+									msg = "정상 가입";
+									
+								} catch (Exception e) {
+									e.printStackTrace();
+									code = 999;
+									msg = "가입 중 오류가 발생하였습니다.";
+								}
+							} else {
+								code = Integer.parseInt(resultCode);
+								msg = resultMsg;
 							}
 						}
 						// 문제가 없다면 정상코드 제공
 						responseJSON.put("code", code);
 						responseJSON.put("msg", msg);
-
+						
+						logger.info("response JSON : {}", responseJSON.toJSONString());
 						return responseJSON.toJSONString();
 					});
 
@@ -189,16 +215,42 @@ public class SubscribeController {
 						
 						List<Subscribe> targets = service.getSubscribeByMobileno(data);
 						if (targets != null && targets.size()>0) {
-							int dcnt = 0;
-							for (Subscribe d: targets) {
-								service.deleteSubscribe(d.getId());
-								dcnt++;
-								logger.info("[{}] Cancel Service - ID:{}", dcnt, d.getId());
-							}
-							res.status(200);
 							
-							code = 200;
-							msg = "정상 처리";
+							// 통신사로 부가서비스 취소 요청
+							
+							// 부가서비스 제공사로 해지 요청
+							String method = "POST";
+							String uri = "/api/v1.0/linksafe/cancel";
+							JSONObject jsonParam = new JSONObject();
+							jsonParam.put("spcode", data.getSpcode());
+							jsonParam.put("mobileno", data.getMobileno());
+							jsonParam.put("status", "C");
+							
+							JSONObject itfJSON = (JSONObject) JSONValue.parse(itfMgr.sendRequest(method, uri, jsonParam.toJSONString()));
+							String resultCode = itfJSON.getAsString("code");
+							String resultMsg = itfJSON.getAsString("msg");
+							// 아무 문제가 없으면 DB에 저장함 
+							
+							if ("200".equals(resultCode)) {
+								try {
+									int dcnt = 0;
+									for (Subscribe d: targets) {
+										service.deleteSubscribe(d.getId());
+										dcnt++;
+										logger.info("[{}] Cancel Service - ID:{}", dcnt, d.getId());
+									}
+									code = 200;
+									msg = "정상 해지";
+									
+								} catch (Exception e) {
+									e.printStackTrace();
+									code = 999;
+									msg = "해지 중 오류가 발생하였습니다.";
+								}
+							} else {
+								code = Integer.parseInt(resultCode);
+								msg = resultMsg;
+							}
 							
 						} else {
 							code = 400;
